Added initial values generator
Creates two folders, one with the initial hash values and the other with a set of constant words used in the hashing algorithm.
The first folder contains the a file for each of the first 8 hash values which are the first 32 bits of the fractional parts of the square roots of the first 8 prime numbers in hexadecimal representation. The second folder contains a file for each of the 64 constant 32-bit words which are the fractional parts of the cube roots of the first 64 primes in hexadecimal representation.